+++ package.json
@@ -13,7 +13,7 @@
     "@testing-library/user-event": "^13.3.0",
     "@types/jest": "^26.0.15",
     "@types/node": "^12.0.0",
-    "@types/react": "^17.0.0",
+    "@types/react": "^16.0.25",
     "@types/react-dom": "^17.0.0",
     "antd": "^4.16.13",
     "autoprefixer": "^9",
