--- src/components/ui/Header.tsx
@@ -1,274 +1,0 @@
-import React from 'react';
-import {
-  fade,
-  makeStyles,
-  Theme,
-  createStyles,
-} from '@material-ui/core/styles';
-import AppBar from '@material-ui/core/AppBar';
-import Toolbar from '@material-ui/core/Toolbar';
-import IconButton from '@material-ui/core/IconButton';
-import Typography from '@material-ui/core/Typography';
-import InputBase from '@material-ui/core/InputBase';
-import Badge from '@material-ui/core/Badge';
-import MenuItem from '@material-ui/core/MenuItem';
-import Menu from '@material-ui/core/Menu';
-import MenuIcon from '@material-ui/icons/Menu';
-import SearchIcon from '@material-ui/icons/Search';
-import AccountCircle from '@material-ui/icons/AccountCircle';
-import MailIcon from '@material-ui/icons/Mail';
-import NotificationsIcon from '@material-ui/icons/Notifications';
-import MoreIcon from '@material-ui/icons/MoreVert';
-import Button from '@material-ui/core/Button';
-import { FaUserTimes } from 'react-icons/fa';
-import { Link } from 'react-router-dom';
-
-const blueColor = '#2DBECD';
-const greenColor = '#51D596';
-const useStyles = makeStyles((theme: Theme) =>
-  createStyles({
-    grow: {
-      flexGrow: 1,
-    },
-    menuButton: {
-      marginRight: theme.spacing(2),
-    },
-    title: {
-      display: 'none',
-      [theme.breakpoints.up('sm')]: {
-        display: 'block',
-      },
-    },
-    search: {
-      // flexGrow:'1',
-
-      position: 'relative',
-      borderRadius: theme.shape.borderRadius,
-      backgroundColor: fade(theme.palette.common.white, 0.15),
-      '&:hover': {
-        backgroundColor: fade(theme.palette.common.white, 0.25),
-      },
-      marginRight: theme.spacing(2),
-      marginLeft: 0,
-      width: '100%',
-      [theme.breakpoints.up('sm')]: {
-        margin: '0 auto',
-        // marginLeft: theme.spacing(3),
-        width: 'auto',
-      },
-    },
-    searchIcon: {
-      padding: theme.spacing(0, 2),
-      height: '100%',
-      position: 'absolute',
-      pointerEvents: 'none',
-      display: 'flex',
-      alignItems: 'center',
-      justifyContent: 'center',
-    },
-    inputRoot: {
-      color: 'inherit',
-    },
-    inputInput: {
-      padding: theme.spacing(1, 1, 1, 0),
-      // vertical padding + font size from searchIcon
-      paddingLeft: `calc(1em + ${theme.spacing(4)}px)`,
-      transition: theme.transitions.create('width'),
-      width: '100%',
-      [theme.breakpoints.up('md')]: {
-        width: '25ch',
-      },
-    },
-    sectionDesktop: {
-      display: 'none',
-      [theme.breakpoints.up('md')]: {
-        display: 'flex',
-        alignItems: 'center',
-      },
-    },
-    headerButtons: {
-      backgroundColor: blueColor,
-      borderRadius: '10px',
-      fontSize: '14px',
-      padding: '6px 8px',
-      textTransform: 'none',
-      marginLeft: '10px',
-      color: '#1B1B25',
-      fontWeight: 700,
-    },
-    headerIcon: {
-      fill: '#1B1B25',
-      width: '20px',
-      // fontSize: '14px',
-      // fontFamily: 'Roboto,sans-serif',
-    },
-    leftHeader: {
-      flexGrow: 1,
-      display: 'flex',
-      alignItems: 'center',
-    },
-    logo: {
-      flexGrow: 0,
-      display: 'flex',
-    },
-    searchFiels: {
-      flexGrow: 0,
-    },
-    sectionMobile: {
-      display: 'flex',
-      [theme.breakpoints.up('md')]: {
-        display: 'none',
-      },
-    },
-    btnText: {
-      padding: '0 0 0 5px ',
-    },
-    headerLink: {
-      margin: '0 20px 0 50px',
-      color: '#fff',
-      fontSize: '20px',
-      textDecoration: 'none',
-      display: 'flex',
-      alignItems: 'center',
-    },
-    toolbar: {
-      backgroundColor: greenColor,
-    },
-  })
-);
-
-export const Header = () => {
-  const classes = useStyles();
-  const [anchorEl, setAnchorEl] = React.useState(null);
-  const [mobileMoreAnchorEl, setMobileMoreAnchorEl] = React.useState(null);
-
-  const isMenuOpen = Boolean(anchorEl);
-  const isMobileMenuOpen = Boolean(mobileMoreAnchorEl);
-
-  const handleProfileMenuOpen = (event: any) => {
-    setAnchorEl(event.currentTarget);
-  };
-
-  const handleMobileMenuClose = () => {
-    setMobileMoreAnchorEl(null);
-  };
-
-  const handleMenuClose = () => {
-    setAnchorEl(null);
-    handleMobileMenuClose();
-  };
-
-  const handleMobileMenuOpen = (event: any) => {
-    setMobileMoreAnchorEl(event.currentTarget);
-  };
-
-  const menuId = 'primary-search-account-menu';
-  // const renderMenu = (
-  //   <Menu
-  //     anchorEl={anchorEl}
-  //     anchorOrigin={{ vertical: 'top', horizontal: 'right' }}
-  //     id={menuId}
-  //     keepMounted
-  //     transformOrigin={{ vertical: 'top', horizontal: 'right' }}
-  //     open={isMenuOpen}
-  //     onClose={handleMenuClose}
-  //   >
-  //     <MenuItem onClick={handleMenuClose}>Profile</MenuItem>
-  //     <MenuItem onClick={handleMenuClose}>My account</MenuItem>
-  //   </Menu>
-  // );
-
-  // const mobileMenuId = 'primary-search-account-menu-mobile';
-  // const renderMobileMenu = (
-  //   <Menu
-  //     anchorEl={mobileMoreAnchorEl}
-  //     anchorOrigin={{ vertical: 'top', horizontal: 'right' }}
-  //     id={mobileMenuId}
-  //     keepMounted
-  //     transformOrigin={{ vertical: 'top', horizontal: 'right' }}
-  //     open={isMobileMenuOpen}
-  //     onClose={handleMobileMenuClose}
-  //   >
-
-  //   </Menu>
-  // );
-
-  return (
-    <div className={classes.grow}>
-      <AppBar position="static">
-        <Toolbar className={classes.toolbar}>
-          <div className={classes.leftHeader}>
-            <div className={classes.logo}>
-              <Typography className={classes.title} variant="h6" noWrap>
-                Fresh Recipes
-              </Typography>
-
-              <Link to="/" className={classes.headerLink}>
-                Главная
-              </Link>
-
-              <Link to="/recipe/new" className={classes.headerLink}>
-                Добавить рецепт
-              </Link>
-            </div>
-            <div className={classes.search}>
-              <div className={classes.searchIcon}>
-                <SearchIcon />
-              </div>
-              <InputBase
-                placeholder="Search…"
-                classes={{
-                  root: classes.inputRoot,
-                  input: classes.inputInput,
-                }}
-                inputProps={{ 'aria-label': 'search' }}
-              />
-            </div>
-          </div>
-          {/* <div className={classes.grow} /> */}
-          <div className={classes.sectionDesktop}>
-            <Button className={classes.headerButtons} color="inherit">
-              <FaUserTimes className={classes.headerIcon} />
-              <Link to="/sign-in">
-                <span className={classes.btnText}>Login</span>
-              </Link>
-            </Button>
-
-            {/* <Button color="inherit">
-
-            </Button> */}
-
-            <IconButton
-              edge="end"
-              aria-label="account of current user"
-              aria-controls={menuId}
-              aria-haspopup="true"
-              // onClick={handleProfileMenuOpen}
-              color="inherit"
-              className={classes.headerButtons}
-            >
-              <AccountCircle className={classes.headerIcon} />
-              <Link to="/sign-up">
-                <span className={classes.btnText}>Sign Up</span>
-              </Link>
-            </IconButton>
-          </div>
-          {/* <div className={classes.sectionMobile}>
-            <IconButton
-              aria-label="show more"
-              // aria-controls={mobileMenuId}
-              aria-haspopup="true"
-              onClick={handleMobileMenuOpen}
-              color="inherit"
-            >
-              <MoreIcon />
-            </IconButton>
-          </div> */}
-        </Toolbar>
-      </AppBar>
-
-      {/* {renderMobileMenu} */}
-      {/* {renderMenu} */}
-    </div>
-  );
-};
