+++ src/components/ui/Header/Header.tsx
@@ -1,0 +1,155 @@
+import React from 'react';
+import AppBar from '@material-ui/core/AppBar';
+import Toolbar from '@material-ui/core/Toolbar';
+import IconButton from '@material-ui/core/IconButton';
+import Typography from '@material-ui/core/Typography';
+import InputBase from '@material-ui/core/InputBase';
+import Badge from '@material-ui/core/Badge';
+import MenuItem from '@material-ui/core/MenuItem';
+import Menu from '@material-ui/core/Menu';
+import MenuIcon from '@material-ui/icons/Menu';
+import SearchIcon from '@material-ui/icons/Search';
+import AccountCircle from '@material-ui/icons/AccountCircle';
+import MailIcon from '@material-ui/icons/Mail';
+import NotificationsIcon from '@material-ui/icons/Notifications';
+import MoreIcon from '@material-ui/icons/MoreVert';
+import Button from '@material-ui/core/Button';
+import { FaUserTimes } from 'react-icons/fa';
+import { Link } from 'react-router-dom';
+import { useStyles } from './styles';
+
+export const Header = () => {
+    const classes = useStyles();
+    const [anchorEl, setAnchorEl] = React.useState(null);
+    const [mobileMoreAnchorEl, setMobileMoreAnchorEl] = React.useState(null);
+
+    const isMenuOpen = Boolean(anchorEl);
+    const isMobileMenuOpen = Boolean(mobileMoreAnchorEl);
+
+    const handleProfileMenuOpen = (event: any) => {
+        setAnchorEl(event.currentTarget);
+    };
+
+    const handleMobileMenuClose = () => {
+        setMobileMoreAnchorEl(null);
+    };
+
+    const handleMenuClose = () => {
+        setAnchorEl(null);
+        handleMobileMenuClose();
+    };
+
+    const handleMobileMenuOpen = (event: any) => {
+        setMobileMoreAnchorEl(event.currentTarget);
+    };
+
+    const menuId = 'primary-search-account-menu';
+    // const renderMenu = (
+    //   <Menu
+    //     anchorEl={anchorEl}
+    //     anchorOrigin={{ vertical: 'top', horizontal: 'right' }}
+    //     id={menuId}
+    //     keepMounted
+    //     transformOrigin={{ vertical: 'top', horizontal: 'right' }}
+    //     open={isMenuOpen}
+    //     onClose={handleMenuClose}
+    //   >
+    //     <MenuItem onClick={handleMenuClose}>Profile</MenuItem>
+    //     <MenuItem onClick={handleMenuClose}>My account</MenuItem>
+    //   </Menu>
+    // );
+
+    // const mobileMenuId = 'primary-search-account-menu-mobile';
+    // const renderMobileMenu = (
+    //   <Menu
+    //     anchorEl={mobileMoreAnchorEl}
+    //     anchorOrigin={{ vertical: 'top', horizontal: 'right' }}
+    //     id={mobileMenuId}
+    //     keepMounted
+    //     transformOrigin={{ vertical: 'top', horizontal: 'right' }}
+    //     open={isMobileMenuOpen}
+    //     onClose={handleMobileMenuClose}
+    //   >
+
+    //   </Menu>
+    // );
+
+    return (
+        <div className={classes.grow}>
+            <AppBar position="static">
+                <Toolbar className={classes.toolbar}>
+                    <div className={classes.leftHeader}>
+                        <div className={classes.logo}>
+                            <Typography className={classes.title} variant="h6" noWrap>
+                                Fresh Recipes
+                            </Typography>
+
+                            <Link to="/" className={classes.headerLink}>
+                                Главная
+                            </Link>
+
+                            <Link to="/recipe/new" className={classes.headerLink}>
+                                Добавить рецепт
+                            </Link>
+                        </div>
+                        <div className={classes.search}>
+                            <div className={classes.searchIcon}>
+                                <SearchIcon />
+                            </div>
+                            <InputBase
+                                placeholder="Search…"
+                                classes={{
+                                    root: classes.inputRoot,
+                                    input: classes.inputInput,
+                                }}
+                                inputProps={{ 'aria-label': 'search' }}
+                            />
+                        </div>
+                    </div>
+                    {/* <div className={classes.grow} /> */}
+                    <div className={classes.sectionDesktop}>
+                        <Button className={classes.headerButtons} color="inherit">
+                            <FaUserTimes className={classes.headerIcon} />
+                            <Link to="/sign-in">
+                                <span className={classes.btnText}>Login</span>
+                            </Link>
+                        </Button>
+
+                        {/* <Button color="inherit">
+
+            </Button> */}
+
+                        <IconButton
+                            edge="end"
+                            aria-label="account of current user"
+                            aria-controls={menuId}
+                            aria-haspopup="true"
+                            // onClick={handleProfileMenuOpen}
+                            color="inherit"
+                            className={classes.headerButtons}
+                        >
+                            <AccountCircle className={classes.headerIcon} />
+                            <Link to="/sign-up">
+                                <span className={classes.btnText}>Sign Up</span>
+                            </Link>
+                        </IconButton>
+                    </div>
+                    {/* <div className={classes.sectionMobile}>
+            <IconButton
+              aria-label="show more"
+              // aria-controls={mobileMenuId}
+              aria-haspopup="true"
+              onClick={handleMobileMenuOpen}
+              color="inherit"
+            >
+              <MoreIcon />
+            </IconButton>
+          </div> */}
+                </Toolbar>
+            </AppBar>
+
+            {/* {renderMobileMenu} */}
+            {/* {renderMenu} */}
+        </div>
+    );
+};
