+++ src/components/ui/Home/styles.tsx
@@ -1,0 +1,9 @@
+import { makeStyles, createStyles } from '@material-ui/core';
+
+export const useStyles = makeStyles((theme?: any) =>
+    createStyles({
+        root: {
+            color: '#000',
+        },
+    })
+);
