+++ src/components/ui/RecipeEdit.tsx
@@ -2,5 +2,5 @@ import React from 'react';
 import { makeStyles, Theme, createStyles } from '@material-ui/core/styles';
 
 export const RecipeEdit = () => {
-  <div className="ololo">11212</div>;
+    <div className="ololo">11212</div>;
 };
