+++ src/components/ui/RecipeNew.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import { fade, makeStyles } from '@material-ui/core/styles';
-import { RecipeNewForm } from './RecipeNewForm';
+import { RecipeNewForm } from './RecipeNewForm/RecipeNewForm';
 
 export const RecipeNew = () => (
-  <div>
-    <RecipeNewForm />
-  </div>
+    <div>
+        <RecipeNewForm />
+    </div>
 );
