+++ src/components/ui/SignIn/styles.tsx
@@ -1,0 +1,19 @@
+import { makeStyles, createStyles } from '@material-ui/core';
+
+export const useStyles = makeStyles((theme?: any) =>
+    createStyles({
+        root: {
+            display: 'flex',
+            flexWrap: 'wrap',
+        },
+        margin: {
+            margin: theme.spacing(1),
+        },
+        withoutLabel: {
+            marginTop: theme.spacing(3),
+        },
+        textField: {
+            width: '25ch',
+        },
+    })
+);
