--- src/components/ui/footer/footer.module.css
@@ -1,7 +1,0 @@
-footer{
-    height: 1px;
-    display: table-row;
-    background-color: #D1D1D1;
-    text-align: center;
-
-}
